+++ R/npdims.R
@@ -1,7 +1,7 @@
 #' @export
 universals::npdims
 
-#' Number of Parameter Dimensions of a term
+#' Number of Dimensions of each Parameter
 #'
 #' @inherit universals::npars
 #' @param terms A flag specifying whether to get the number of dimensions for each term element.
@@ -9,18 +9,16 @@ universals::npdims
 #'
 #' @examples
 #' npdims(as.term(c("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]")))
-#' npdims(as.term(c("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]")), terms = TRUE)
 npdims.term <- function(x, terms = FALSE, ...) {
   chk_flag(terms)
   chk_unused(...)
-
-  if (!terms) {
-    return(vapply(pdims(x), length, 1L))
+  
+  if(!missing(terms)) {
+    deprecate_soft("0.1.0.9003", "npdims(terms =)", details = "If `terms = TRUE` use `npdims_terms() otherwise replace `npdims(terms = FALSE)` with `npdims()`.")
   }
-  x <- tindex(x)
-  names(x) <- pars_terms(as.term(names(x)))
-  is.na <- vapply(x, identical, TRUE, y = NA_integer_)
-  x <- vapply(x, length, 1L)
-  is.na(x[is.na]) <- TRUE
-  x
+  
+  if (terms) {
+    return(npdims_terms(x))
+  }
+  vapply(pdims(x), length, 1L)
 }