+++ R/tindex.R
@@ -1,21 +1,20 @@
-#' @export
-universals::tindex
-
 #' Term Index
 #'
 #' Gets the index for each term of an object.
 #'
 #' For example the index of `beta[2,1]` is `c(2L, 1L)`
 #' while the index for `sigma` is `1L`.
-#' 
 #' It is useful for extracting the values of individual terms.
 #'
-#' @inherit universals::tindex
+#' @inheritParams params
+#' @return A named list of integer vectors of the index for each term.
+#' @family {dimensions}
 #' @export
 #'
 #' @examples
 #' tindex(as.term(c("alpha", "alpha[2]", "beta[1,1]", "beta[2 ,1  ]")))
-tindex.term <- function(x, ...) {
+tindex <- function(x) {
+  x <- as.term(x)
   names <- x
   x <- sub(p0("^", .par_name_pattern), "", x)
   x <- sub("^$", "1", x)
@@ -26,9 +25,4 @@ tindex.term <- function(x, ...) {
   x[is.na(names)] <- NA_integer_
   names(x) <- names
   x
-}
-
-# internal use only
-tindex.character <- function(x, ...) {
-  tindex(as.term(x))
 }